+++ NEWS.md
@@ -3,10 +3,13 @@
 
 * The development branch for `renv` has moved from master to main.
 
+* `renv::init()` gains the `bioconductor` argument, to be used to initialize
+  a project with a particular Bioconductor release.
+
 * Project settings can now be overridden by setting an R option of the same
   name. For example, one could force a project to use Bioconductor 3.14 by
   setting `options(renv.settings.bioconductor.version = "3.14")` within
-  the project `.Rprofile` (or similar startup profile).
+  the project `.Rprofile` (or similar startup R profile).
 
 * The ad-hoc package repository called "local sources" has been renamed to
   the "package cellar". In addition, the path to the cellar is now